make IDs of type long

+++ csc1106/src/main/java/webprogramming/csc1106/Entities/Post.java
@@ -16,7 +16,7 @@ public class Post {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id; // primary key
+    private long id;
 
     @ManyToOne
     private Category category;
